Modify architecture and apply clean archi
We were calling 3 usecase from the controller, or this goes against the
clean architecture use in Pix.
Now we call 1 usecase squashing the logic previously done in the
controller.
