+++ api/lib/domain/usecases/index.js
@@ -57,9 +57,7 @@ function injectDependencies(usecases) {
 module.exports = injectDependencies({
   acceptPixCertifTermsOfService: require('./accept-pix-certif-terms-of-service'),
   acceptPixOrgaTermsOfService: require('./accept-pix-orga-terms-of-service'),
-  addOrganizationLogoToCampaign: require('./add-organization-logo-to-campaign'),
   addOrganizationMembershipWithEmail: require('./add-organization-membership-with-email'),
-  assertUserBelongToOrganization: require('./assert-user-belong-to-organization'),
   authenticateUser: require('./authenticate-user'),
   beginCampaignParticipationImprovement: require('./begin-campaign-participation-improvement'),
   completeAssessment: require('./complete-assessment'),
